+++ vclib-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciProcessTest.kt
@@ -6,11 +6,13 @@ import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.VerifiableCredentialJws
 import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
 import at.asitplus.wallet.lib.iso.IssuerSigned
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements
 import at.asitplus.wallet.lib.jws.JwsSigned
 import at.asitplus.wallet.lib.oidc.DummyCredentialDataProvider
-import at.asitplus.wallet.lib.oidc.OpenIdConstants
 import at.asitplus.wallet.lib.oidc.OpenIdConstants.GRANT_TYPE_CODE
+import at.asitplus.wallet.lib.oidc.OpenIdConstants.TOKEN_PREFIX_BEARER
 import io.kotest.core.spec.style.FunSpec
+import io.kotest.matchers.ints.shouldBeGreaterThan
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.ktor.http.Url
@@ -35,10 +37,10 @@ class OidvciProcessTest : FunSpec({
         )
         val credential = runProcess(issuer, client)
         credential.format shouldBe CredentialFormatEnum.JWT_VC
-        credential.credential.shouldNotBeNull()
-        println(credential.credential)
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
 
-        val jws = JwsSigned.parse(credential.credential!!)
+        val jws = JwsSigned.parse(serializedCredential)
         jws.shouldNotBeNull()
         val vcJws = VerifiableCredentialJws.deserialize(jws.payload.decodeToString())
         vcJws.shouldNotBeNull().also { println(it) }
@@ -51,13 +53,32 @@ class OidvciProcessTest : FunSpec({
         )
         val credential = runProcess(issuer, client)
         credential.format shouldBe CredentialFormatEnum.JWT_VC_SD
-        credential.credential.shouldNotBeNull()
-        println(credential.credential)
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
 
-        val jws = JwsSigned.parse(credential.credential!!)
+        val jws = JwsSigned.parse(serializedCredential)
         jws.shouldNotBeNull()
         val sdJwt = VerifiableCredentialSdJwt.deserialize(jws.payload.decodeToString())
         sdJwt.shouldNotBeNull().also { println(it) }
+        sdJwt.disclosureDigests.size shouldBeGreaterThan 1
+    }
+
+    test("process with W3C VC SD-JWT one requested claim") {
+        val client = WalletService(
+            credentialScheme = ConstantIndex.AtomicAttribute2023,
+            credentialRepresentation = ConstantIndex.CredentialRepresentation.SD_JWT,
+            requestedAttributes = listOf("family-name")
+        )
+        val credential = runProcess(issuer, client)
+        credential.format shouldBe CredentialFormatEnum.JWT_VC_SD
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
+
+        val jws = JwsSigned.parse(serializedCredential)
+        jws.shouldNotBeNull()
+        val sdJwt = VerifiableCredentialSdJwt.deserialize(jws.payload.decodeToString())
+        sdJwt.shouldNotBeNull().also { println(it) }
+        sdJwt.disclosureDigests.size shouldBe 1
     }
 
     test("process with ISO mobile driving licence") {
@@ -67,10 +88,32 @@ class OidvciProcessTest : FunSpec({
         )
         val credential = runProcess(issuer, client)
         credential.format shouldBe CredentialFormatEnum.MSO_MDOC
-        credential.credential.shouldNotBeNull()
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
 
-        val issuerSigned = IssuerSigned.deserialize(credential.credential!!.decodeToByteArray(Base64()))
+        val issuerSigned = IssuerSigned.deserialize(serializedCredential.decodeToByteArray(Base64()))
         issuerSigned.shouldNotBeNull().also { println(it) }
+        val numberOfClaims = issuerSigned.namespaces?.values?.firstOrNull()?.entries?.size
+        numberOfClaims.shouldNotBeNull()
+        numberOfClaims shouldBeGreaterThan 1
+    }
+
+    test("process with ISO mobile driving licence one requested claim") {
+        val client = WalletService(
+            credentialScheme = ConstantIndex.MobileDrivingLicence2023,
+            credentialRepresentation = ConstantIndex.CredentialRepresentation.ISO_MDOC,
+            requestedAttributes = listOf(MobileDrivingLicenceDataElements.DOCUMENT_NUMBER)
+        )
+        val credential = runProcess(issuer, client)
+        credential.format shouldBe CredentialFormatEnum.MSO_MDOC
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
+
+        val issuerSigned = IssuerSigned.deserialize(serializedCredential.decodeToByteArray(Base64()))
+        issuerSigned.shouldNotBeNull().also { println(it) }
+        val numberOfClaims = issuerSigned.namespaces?.values?.firstOrNull()?.entries?.size
+        numberOfClaims.shouldNotBeNull()
+        numberOfClaims shouldBe 1
     }
 
     test("process with ISO atomic attributes") {
@@ -80,9 +123,10 @@ class OidvciProcessTest : FunSpec({
         )
         val credential = runProcess(issuer, client)
         credential.format shouldBe CredentialFormatEnum.MSO_MDOC
-        credential.credential.shouldNotBeNull()
+        val serializedCredential = credential.credential
+        serializedCredential.shouldNotBeNull().also { println(it) }
 
-        val issuerSigned = IssuerSigned.deserialize(credential.credential!!.decodeToByteArray(Base64()))
+        val issuerSigned = IssuerSigned.deserialize(serializedCredential.decodeToByteArray(Base64()))
         issuerSigned.shouldNotBeNull().also { println(it) }
     }
 
@@ -101,5 +145,5 @@ private suspend fun runProcess(
     val tokenRequest = client.createTokenRequestParameters(code)
     val token = issuer.token(tokenRequest)
     val credentialRequest = client.createCredentialRequest(token, metadata)
-    return issuer.credential(OpenIdConstants.TOKEN_PREFIX_BEARER + token.accessToken, credentialRequest)
+    return issuer.credential(TOKEN_PREFIX_BEARER + token.accessToken, credentialRequest)
 }