+++ vclib/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/DummyCredentialDataProvider.kt
@@ -7,6 +7,7 @@ import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.iso.ElementValue
 import at.asitplus.wallet.lib.iso.IssuerSignedItem
 import kotlinx.datetime.Clock
+import kotlinx.datetime.LocalDate
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.minutes
 
@@ -19,7 +20,8 @@ class DummyCredentialDataProvider(
     override fun getCredential(
         subjectPublicKey: CryptoPublicKey,
         credentialScheme: ConstantIndex.CredentialScheme,
-        representation: ConstantIndex.CredentialRepresentation
+        representation: ConstantIndex.CredentialRepresentation,
+        claimNames: Collection<String>?
     ): KmmResult<List<CredentialToBeIssued>> {
         val expiration = clock.now() + defaultLifetime
         val claims = listOf(
@@ -38,7 +40,7 @@ class DummyCredentialDataProvider(
 
             ConstantIndex.CredentialRepresentation.PLAIN_JWT -> claims.map { claim ->
                 CredentialToBeIssued.VcJwt(
-                    subject = AtomicAttribute2023(subjectId, claim.name, claim.value),
+                    subject = AtomicAttribute2023(subjectId, claim.name, claim.value.toString()),
                     expiration = expiration,
                 )
             } + CredentialToBeIssued.VcJwt(
@@ -50,7 +52,7 @@ class DummyCredentialDataProvider(
             ConstantIndex.CredentialRepresentation.ISO_MDOC -> listOf(
                 CredentialToBeIssued.Iso(
                     issuerSignedItems = claims.mapIndexed { index, claim ->
-                        buildIssuerSignedItem(claim.name, claim.value, index.toUInt())
+                        issuerSignedItem(claim.name, claim.value, index.toUInt())
                     },
                     expiration = expiration,
                 )
@@ -59,11 +61,17 @@ class DummyCredentialDataProvider(
         return KmmResult.success(credentials)
     }
 
-    private fun buildIssuerSignedItem(elementIdentifier: String, elementValue: String, digestId: UInt) =
+    private fun issuerSignedItem(name: String, value: Any, digestId: UInt) =
         IssuerSignedItem(
             digestId = digestId,
             random = Random.nextBytes(16),
-            elementIdentifier = elementIdentifier,
-            elementValue = ElementValue(string = elementValue)
+            elementIdentifier = name,
+            elementValue = when (value) {
+                is String -> ElementValue(string = value)
+                is ByteArray -> ElementValue(bytes = value)
+                is LocalDate -> ElementValue(date = value)
+                is Boolean -> ElementValue(boolean = value)
+                else -> ElementValue(string = value.toString())
+            }
         )
 }